+++ src/app/components/endre-periode-form-renderer/EndrePeriodeFormRenderer.tsx
@@ -1,0 +1,95 @@
+import * as React from 'react';
+import {
+    Periode,
+    Periodetype,
+    Utsettelsesperiode,
+    Uttaksperiode,
+    Oppholdsperiode,
+    Overføringsperiode
+} from '../../types/uttaksplan/periodetyper';
+import BEMHelper from 'common/util/bem';
+import { preventFormSubmit } from 'common/util/eventUtils';
+import { RecursivePartial } from '../../types/Partial';
+import { FormattedMessage, injectIntl, InjectedIntlProps } from 'react-intl';
+import { connect } from 'react-redux';
+import { DispatchProps } from 'common/redux/types';
+import søknadActionCreators from '../../redux/actions/søknad/søknadActionCreators';
+import BekreftDialog from 'common/components/dialog/BekreftDialog';
+import getMessage from 'common/util/i18nUtils';
+import { cleanupPeriode } from '../../util/cleanup/periodeCleanup';
+
+export interface OwnProps {
+    periode: Periode;
+    render: (onChange: (periode: RecursivePartial<Periode>) => void, onRequestDelete: () => void) => {};
+}
+
+interface State {
+    visBekreftSlettDialog: boolean;
+}
+const bem = BEMHelper('endrePeriodeForm');
+
+type Props = OwnProps & DispatchProps & InjectedIntlProps;
+
+class EndrePeriodeFormRenderer extends React.Component<Props, State> {
+    constructor(props: Props) {
+        super(props);
+        this.onChange = this.onChange.bind(this);
+        this.onRequestDelete = this.onRequestDelete.bind(this);
+        this.onCancelDelete = this.onCancelDelete.bind(this);
+        this.onDelete = this.onDelete.bind(this);
+        this.state = {
+            visBekreftSlettDialog: false
+        };
+    }
+    onChange(p: RecursivePartial<Periode>) {
+        let updatedPeriode: Periode | undefined;
+        const { periode, dispatch } = this.props;
+        if (periode.type === Periodetype.Utsettelse) {
+            updatedPeriode = { ...periode, ...(p as Utsettelsesperiode) };
+        } else if (periode.type === Periodetype.Uttak) {
+            updatedPeriode = { ...periode, ...(p as Uttaksperiode) };
+        } else if (periode.type === Periodetype.Overføring) {
+            updatedPeriode = {
+                ...periode,
+                ...(p as Overføringsperiode)
+            };
+        } else if (periode.type === Periodetype.Opphold) {
+            updatedPeriode = { ...periode, ...(p as Oppholdsperiode) };
+        }
+        if (updatedPeriode !== undefined) {
+            if (updatedPeriode.type !== this.props.periode.type) {
+                updatedPeriode.vedlegg = [];
+            }
+            dispatch(søknadActionCreators.uttaksplanUpdatePeriode(cleanupPeriode(updatedPeriode)));
+        }
+    }
+    onDelete() {
+        this.props.dispatch(søknadActionCreators.uttaksplanDeletePeriode(this.props.periode));
+    }
+    onRequestDelete() {
+        this.setState({ visBekreftSlettDialog: true });
+    }
+    onCancelDelete() {
+        this.setState({ visBekreftSlettDialog: false });
+    }
+    render() {
+        const { periode, render, intl } = this.props;
+        return (
+            <form className={bem.className} onSubmit={preventFormSubmit}>
+                {render(this.onChange, this.onRequestDelete)}
+                <BekreftDialog
+                    tittel={getMessage(intl, `endrePeriodeForm.bekreftSlettDialog.${periode.type}.tittel`)}
+                    contentLabel={getMessage(intl, `endrePeriodeForm.bekreftSlettDialog.${periode.type}.tittel`)}
+                    isOpen={this.state.visBekreftSlettDialog}
+                    onBekreft={this.onDelete}
+                    onRequestClose={this.onCancelDelete}
+                    bekreftLabel={getMessage(intl, 'ja')}
+                    avbrytLabel={getMessage(intl, 'nei')}>
+                    <FormattedMessage id="endrePeriodeForm.bekreftSlettDialog.uttak.melding" />
+                </BekreftDialog>
+            </form>
+        );
+    }
+}
+
+export default connect()(injectIntl(EndrePeriodeFormRenderer));
